Remove vfsStream dependency
Although I can not find the original story in which we added it, it was added as a way to make sure that PHPUnit or Behat tests didn't fail. When we added it it was silently required by Drupal tests but not yet added. We now keep updating it but we should just get it from [Drupal's dev dependencies](https://github.com/drupal/core-dev/blob/9.0.x/composer.json#L17) since as far as I know we don't use it directly.