Remove drupal/core-dev as dependency
The dependency is created specifically to run the Drupal core tests,
which is not actually something we want to do.

All the dependencies we care about for our own infrastructure are
already part of open_social_dev itself, which lets us control them in a
way that also allows us to do our Open Social upgrade paths.