Implement distribution compatibility checking
Since we use our development tools for our upgrade path testing we want
to make sure that changes to the tools don't accidentally break our
downstream CIs.

It's too much effort to run entire test suites but we can at least
ensure that composer compatibility works for the current and previous
majore version of the distribution to help us catch issues early.

We ignore platform requirements because we just want to make sure that
there's a compatible set of dependencies. We'll leave the actual PHP
requirements to the downstream projects to manage (and test).