Ignore dev patch included in Open Social 10
Open Social 10 contains a patch for one of our dev dependencies that is
actually outdated.

Unfortunately we must ignore in our `test/` composer.json and can't do
this in our _dev package itself as per
https://github.com/cweagans/composer-patches/issues/112 so it's only a
fix for our own CI and Open Social will have to fix its own stuff in not
patching dev dependencies (but updating this package instead).