Add Drupal test dependencies
The base test classes do need some dependencies which are found in
drupal/core-dev so we add them manually.

The default PHPUnit configuration seems to use the bridge to warn for
deprecations in Drupal core.

We don't care about a specific version because we're happy with whatever
Symfony/Drupal indicates is compatible with our other selected
dependencies.

vfsStream is also needed for some filesystem operations.