Add Behat linting to our toolchain
[Cucumber Linter](github.com/Kingdutch/cucumber-linter) provides a
package that can lint for certain patterns in Behat tests (such as
proper indentation and order of `arrange`, `act`, and `assert` test
stages. This ensures tests are written in a uniform manner making them
easy for others to understand. Additionally duplication of test stages
usually indicates a test should be split up.