Add jangregor/phpstan-prophecy as dev dependency
The phpspec/prophecy tool was adopted for mocking in PHPUnit classes
(though it's not entirely clear why this tool over PHPUnit's built-in
tools). To make sure PHPStan understands what the tool does we need an
extension that tells PHPStan it's still the underlying mocked class. The
added extension is such an extension and was previously added to the
distribution only, but is now added to our dev package so we can use it
everywhere.